refactor/unify Twig extensions + runtimes
