--- core-bundle/src/Twig/Extension/ImageExtension.php
@@ -1,36 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of Contao.
- *
- * (c) Leo Feyer
- *
- * @license LGPL-3.0-or-later
- */
-
-namespace Contao\CoreBundle\Twig\Extension;
-
-use Contao\CoreBundle\Twig\Runtime\FigureRendererRuntime;
-use Contao\CoreBundle\Twig\Runtime\PictureConfigurationRuntime;
-use Twig\Extension\AbstractExtension;
-use Twig\TwigFunction;
-
-class ImageExtension extends AbstractExtension
-{
-    public function getFunctions(): array
-    {
-        return [
-            new TwigFunction(
-                'contao_figure',
-                [FigureRendererRuntime::class, 'render'],
-                ['is_safe' => ['html']]
-            ),
-            new TwigFunction(
-                'picture_config',
-                [PictureConfigurationRuntime::class, 'fromArray']
-            ),
-        ];
-    }
-}
