+++ core-bundle/src/Twig/Runtime/PictureConfigurationRuntime.php
@@ -26,6 +26,9 @@ final class PictureConfigurationRuntime implements RuntimeExtensionInterface
      */
     private $propertyAccessor;
 
+    /**
+     * @internal
+     */
     public function __construct()
     {
         $this->propertyAccessor = PropertyAccess::createPropertyAccessor();
